Scrapers/requirements.txt
firebase-admin==6.5.0
python-dotenv==1.0.0
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
praw==7.7.1
